requirements.txt
lockfile==0.12.2
luigi==3.6.0
numpy==1.26.4
pandas==2.2.1
psycopg2-binary==2.9.9
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.28
tenacity==8.5.0
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
